Initial commit of code used to generate results in the paper
This is the code used to generate results, with file paths removed.  I intend to add more generalisable functions for the key components of the model when I have time. Currently the majority of the code is a use-case specific wrapper and data processing, which can be ignored. The core model functions are relatively simple to extract and use with other data sources.